s02e02/requirements.txt
python-dotenv
requests
langchain
langchain-openai
Pillow
